+++ Source/FunPlusData.swift
@@ -182,7 +182,6 @@ public class FunPlusData: SessionStatusChangeListener {
         - parameter transactionId:      The unique transaction ID sent back by the payment processor.
         - parameter paymentProcessor:   The payment processor.
         - parameter itemsReceived:      An array string consisting of one or more items received.
-        - parameter currencyReceived:   An array string consisting of one or more types of currency received.
      */
     public func tracePayment(
         amount: Double,
@@ -193,22 +192,20 @@ public class FunPlusData: SessionStatusChangeListener {
         transactionId: String,
         paymentProcessor: String,
         itemsReceived: String,
-        currencyReceived: String,
-        currencyReceivedType: String)
+        currencyReceived: String)
     {
         let event = buildDataEvent(
             eventName: "payment",
             customProperties: [
-                "amount":                   String(amount),
-                "currency":                 currency,
-                "iap_product_id":           productId,
-                "iap_product_name":         productName ?? "",
-                "iab_product_type":         productType ?? "",
-                "transaction_id":           transactionId,
-                "payment_processor":        paymentProcessor,
-                "c_items_received":         itemsReceived.toJsonObject() ?? [],
-                "c_currency_received":      currencyReceived.toJsonObject() ?? [],
-                "d_currency_received_type": currencyReceivedType
+                "amount":               String(amount),
+                "currency":             currency,
+                "iap_product_id":       productId,
+                "iap_product_name":     productName ?? "",
+                "iap_product_type":     productType ?? "",
+                "transaction_id":       transactionId,
+                "payment_processor":    paymentProcessor,
+                "c_items_received":     itemsReceived.toJsonObject() ?? [],
+                "c_currency_received":  currencyReceived.toJsonObject() ?? []
             ]
         )